requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.12,<4.0
python-dotenv>=0.19,<1.0
psycopg2>=2.9,<3.0
celery>=5.2,<6.0
redis>=4.0.0
django-environ>=0.4.5,<1.0
django-cors-headers>=3.10,<4.0
gunicorn>=20.1,<21.0
python-decouple==3.6
djangorestframework-simplejwt>=5.0,<6.0